+++ src/main/kotlin/com/dh/baro/look/application/FitFacade.kt
@@ -1,22 +1,25 @@
 package com.dh.baro.look.application
 
+import com.dh.baro.look.application.dto.AiFittingInfo
 import com.dh.baro.look.application.dto.FittingSourceImageUploadInfo
 import com.dh.baro.look.domain.FittingSourceImage
 import com.dh.baro.look.domain.service.FittingSourceImageService
-import com.dh.baro.look.infra.s3.S3ImageClient
+import com.dh.baro.look.infra.gemini.GeminiImageApi
+import com.dh.baro.look.infra.s3.S3ImageApi
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
 @Service
-class FittingSourceImageFacade(
+class FitFacade(
     private val fittingSourceImageService: FittingSourceImageService,
-    private val s3ImageClient: S3ImageClient,
+    private val s3ImageApi: S3ImageApi,
+    private val geminiImageApi: GeminiImageApi,
 ) {
 
     @Transactional
     fun generateUploadUrl(userId: Long): FittingSourceImageUploadInfo {
         val pendingImage = fittingSourceImageService.createPendingImage(userId)
-        val s3Info = s3ImageClient.generatePresignedUrl(pendingImage.id)
+        val s3Info = s3ImageApi.generatePresignedUrl(pendingImage.id)
 
         pendingImage.setS3Key(s3Info.s3Key)
 
@@ -32,12 +35,22 @@ class FittingSourceImageFacade(
     @Transactional
     fun completeImageUpload(imageId: Long, userId: Long) {
         val image = fittingSourceImageService.getFittingSourceImage(imageId, userId)
-        val imageUrl = s3ImageClient.getImageUrl(image.getS3Key()!!)
+        val imageUrl = s3ImageApi.getImageUrl(image.getS3Key()!!)
         fittingSourceImageService.completeImageUpload(image, imageUrl)
     }
 
     @Transactional(readOnly = true)
     fun getUserFittingSourceImages(userId: Long): List<FittingSourceImage> {
         return fittingSourceImageService.getUserFittingSourceImages(userId)
+    }
+
+    fun generateAiFitting(sourceImageUrl: String, clothingImageUrl: String): AiFittingInfo {
+        val generatedImageData = geminiImageApi.generateAiFitting(sourceImageUrl, clothingImageUrl)
+
+        return AiFittingInfo(
+            sourceImageUrl = sourceImageUrl,
+            clothingImageUrl = clothingImageUrl,
+            generatedImageData = generatedImageData,
+        )
     }
 }
